requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
moondream==1.0.3
openai==2.3.0
python-dotenv==1.0.1
pydantic==2.12.1
pydantic_core==2.41.3
sniffio==1.3.1
tqdm==4.67.1
torchvision==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
